+++ src/main/java/org/jeduardo/entries/data/EntryRepository.java
@@ -3,6 +3,6 @@ package org.jeduardo.entries.data;
 import org.jeduardo.entries.model.Entry;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EntryRepository extends CrudRepository<Entry, Long> {
+public interface EntryRepository extends CrudRepository<Entry, Integer> {
     
 }
